+++ src/main/java/de/maxhenkel/camerautils/config/ClientConfig.java
@@ -14,8 +14,8 @@ public class ClientConfig {
     public final ConfigEntry<Double> cinematicCameraModifier;
     public final ConfigEntry<Double> zoom;
     public final ConfigEntry<Double> zoomSensitivity;
-    public final ConfigEntry<Double> thirdPersonDistance;
-    public final ConfigEntry<Double> thirdPersonDistanceSensitivity;
+    public final ConfigEntry<Double> thirdPersonOffset;
+    public final ConfigEntry<Double> thirdPersonOffsetSensitivity;
     public final ConfigEntry<Double> thirdPersonOffsetX1;
     public final ConfigEntry<Double> thirdPersonOffsetY1;
     public final ConfigEntry<Double> thirdPersonOffsetZ1;
@@ -48,8 +48,8 @@ public class ClientConfig {
         cinematicCameraModifier = builder.doubleEntry("cinematic_camera_modifier", 1D, 0D, 1D);
         zoom = builder.doubleEntry("zoom", 0.1D, 0D, 2D);
         zoomSensitivity = builder.doubleEntry("zoom_sensitivity", 0.01D, 0.001D, 1D);
-        thirdPersonDistance = builder.doubleEntry("third_person_zoom", 4D, 0D, 100D);
-        thirdPersonDistanceSensitivity = builder.doubleEntry("third_person_zoom_sensitivity", 0.1D, 0.001D, 1D);
+        thirdPersonOffset = builder.doubleEntry("third_person_offset", 0D, -32D, 32D);
+        thirdPersonOffsetSensitivity = builder.doubleEntry("third_person_offset_sensitivity", 0.1D, 0.001D, 1D);
         thirdPersonOffsetX1 = builder.doubleEntry("third_person_cam_1_offset_x", -4D, -100D, 100D);
         thirdPersonOffsetY1 = builder.doubleEntry("third_person_cam_1_offset_y", 0D, -100D, 100D);
         thirdPersonOffsetZ1 = builder.doubleEntry("third_person_cam_1_offset_z", 0D, -100D, 100D);